+++ src/components/EndpointSetup.js
@@ -1,15 +1,15 @@
 import _ from 'lodash';
 import React from 'react';
-import {AddressParameter} from './ParametersFormComponents/AddressParameter';
-import {AmountParameter} from './ParametersFormComponents/AmountParameter';
-import {AssetCodeParameter} from './ParametersFormComponents/AssetCodeParameter';
-import {AssetTypeParameter} from './ParametersFormComponents/AssetTypeParameter';
-import {CursorParameter} from './ParametersFormComponents/CursorParameter';
-import {LedgerParameter} from './ParametersFormComponents/LedgerParameter';
-import {LimitParameter} from './ParametersFormComponents/LimitParameter';
-import {OperationParameter} from './ParametersFormComponents/OperationParameter';
-import {OrderParameter} from './ParametersFormComponents/OrderParameter';
-import {TransactionParameter} from './ParametersFormComponents/TransactionParameter';
+import {AddressParameter} from './FormComponents/AddressParameter';
+import {AmountParameter} from './FormComponents/AmountParameter';
+import {AssetCodeParameter} from './FormComponents/AssetCodeParameter';
+import {AssetTypeParameter} from './FormComponents/AssetTypeParameter';
+import {CursorParameter} from './FormComponents/CursorParameter';
+import {LedgerParameter} from './FormComponents/LedgerParameter';
+import {LimitParameter} from './FormComponents/LimitParameter';
+import {OperationParameter} from './FormComponents/OperationParameter';
+import {OrderParameter} from './FormComponents/OrderParameter';
+import {TransactionParameter} from './FormComponents/TransactionParameter';
 import {ExplorerActions} from '../actions/ExplorerActions';
 import {ExplorerStore} from '../stores/ExplorerStore';
 import {EasySelect} from './EasySelect';
