+++ src/components/TransactionBuilder.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {EndpointPicker} from './EndpointPicker';
 import {EndpointSetup} from './EndpointSetup';
 import {EndpointResult} from './EndpointResult';
-import {AddressParameter} from './ParametersFormComponents/AddressParameter';
+import {AddressParameter} from './FormComponents/AddressParameter';
 import ImportXDR from './ImportXDR';
 import TxBuilderAttributes from './TxBuilderAttributes';
 import TxBuilderOperations from './TxBuilderOperations';
