+++ src/components/TxBuilderResult.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {AddressParameter} from './ParametersFormComponents/AddressParameter';
+import {AddressParameter} from './FormComponents/AddressParameter';
 import TxBuilderStore from '../stores/TxBuilderStore';
 
 export default class TxBuilderResult extends React.Component {
